Add --bestof option
Can be used any number, the random will be based in this number. For
example, "--bestof 4" we get the first four items of the search result,
then use a random item of these four items.

This is good for a more accurate search, for example `gh gif 3 -R
"batman dancing" -B 3` I already know that only the first 3 items are
cool, but I don't like the others.